pyproject.toml
[project]
name = "awesome-babushka-auth-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "aiosmtplib==4.0.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "awesome-babushka-commons",
    "bandit==1.8.3",
    "bcrypt==4.3.0",
    "certifi==2025.7.14",
    "cfgv==3.4.0",
    "charset-normalizer==3.4.2",
    "click==8.1.8",
    "crypto==1.4.1",
    "deprecated==1.2.18",
    "distlib==0.3.9",
    "dnspython==2.7.0",
    "ecdsa==0.19.0",
    "fastapi==0.116.1",
    "filelock==3.17.0",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "identify==2.6.7",
    "idna==3.10",
    "limits==4.0.1",
    "markdown-it-py==3.0.0",
    "mdurl==0.1.2",
    "naked==0.1.32",
    "nodeenv==1.9.1",
    "packaging==24.2",
    "passlib==1.7.4",
    "pbr==6.1.1",
    "platformdirs==4.3.6",
    "pre-commit==4.1.0",
    "pyasn1==0.6.1",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pygments==2.19.1",
    "pymongo==4.13.2",
    "python-jose==3.3.0",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "requests==2.32.4",
    "rich==14.0.0",
    "rsa==4.9",
    "setuptools==78.1.0",
    "shellescape==3.8.1",
    "six==1.17.0",
    "slowapi==0.1.9",
    "sniffio==1.3.1",
    "starlette==0.47.2",
    "stevedore==5.4.1",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "urllib3==2.5.0",
    "uvicorn==0.34.0",
    "virtualenv==20.29.2",
    "wrapt==1.17.2",
]

[tool.uv.sources]
awesome-babushka-commons = { git = "https://github.com/himansu9805/awesome-babushka-commons.git", rev = "main" }

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100